yeah

+++ database/migrations/2024_01_07_145243_create_items_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->unsignedBigInteger('category_id');
             $table->foreign('category_id')->references('id')->on('categories');
             $table->string("name");
-            $table->integer("price");
-            $table->integer("quantity");
+            $table->bigInteger("price");
+            $table->bigInteger("quantity");
             $table->string("image");
             $table->timestamps();
         });
