+++ database/migrations/2024_02_11_075632_create_orders_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('address');
             $table->string('postal');
-            $table->integer('total');
+            $table->bigInteger('total');
             $table->timestamps();
         });
     }
